move blog.spec to sunset
